Fix references to figures
In-line references to figures used to refer to the wrong ones when compiled.
(e.g., it would refer to Figure 6 when actually the figure in question was 10)

This was fixed by changing the placement of the \label command within
the figure environment: it's now after the \caption command.
This matters because of how the \label macro is defined.
See: https://tex.stackexchange.com/a/111281