Setup and teardown.
The setup scripts generate a "runid" based on the computer name and the
current time and write it to a file. The teardown script stops VLC,
copies the video file and the instrumentation log, and copies them to a
temporary directory, giving them names that include the runid. Then the
teardown script calls s3-uploader.exe, which should take care of
uploading the files from the temporary upload directory.