Be more careful about switching to not_running.
When the user starts the program a second time after a first successful
bootstrapping, the main configuration interface isn't shown; only the
progress bar. When the progress bar fills, it gets an unload event, but
we never get an unload event for the main configuration interface
(because it was never loaded).