Handle better the failed_to_save_settings error case.
Now we keep a copy of the most recently saved values (values that, when
"Connect" was clicked, led to the progress bar being shown). When
starting afresh, load the most recently saved values. When there's a
failed_to_save_settings error, don't save the values.

The previous way we handled failed_to_save_settings (clearing all the
proxy variables) was wrong in two ways: the problem could have been
related to the bridge settings, not the proxy settings; and the
in-memory values persist until you quit the program. So if someone
entered a bad proxy port, say, the interface wouldn't clear away the
proxy address; it would allow them to edit the port and try again.