Remove duplicate submissions from summative-demographics.csv.
This is the result of a manual check by Linda and me.

This entry:
	3/28/16 14:36,3/28/16 14:37,x25,Female,20,College/Trade School
looks like a duplicate of
	3/28/16 14:21,3/28/16 14:22,X25,Female,20,College/Trade School

This entry:
	3/30/16 17:07,3/30/16 17:07,X11,Female,31,Graduate/Advanced Education
looks like a duplicate of
	3/30/16 16:49,3/30/16 16:49,X11,Female,31,Graduate/Advanced Education

This entry (notice bad seat number "3130"):
	3/30/16 19:07,3/30/16 19:09,3130,Female,31,College/Trade School
looks like a duplicate of
	3/30/16 19:16,3/30/16 19:17,X04,Female,31,College/Trade School

With these three deletions, all of the entries in
summative-demographics.csv correspond to exactly one entry in
participants.csv. participants.csv is left with two entries that lack
demographic information from summative-demographics.csv:
	20160323-1400 X23 (P93, good)
	20160328-1400,X06 (P124, not good)