Show total participant count at the beginning of info.R.
The first summary of participants counts was meant to include "bad"
participants:
	number of participants: 124 (114 good, 10 bad)
The list was being filtered too early, resulting in:
	number of participants: 114 (114 good, 0 bad)
The list is re-read after this first line is printed, omitting all "bad"
participants, so this doesn't affect any other lines.

This commit undoes one line of a396064a.